+++ Sources/LocationTracker+iOS.swift
@@ -19,28 +19,31 @@ extension LocationTracker {
     /// The standard location service is most appropriate for apps that deliver location-related
     /// information directly to the user but it may be used by other types of apps too.
     ///
-    /// Never specify an accuracy value greater than what you need.  Core Location uses the accuracy
-    /// value you specify to manage power better.  A higher degree of accuracy requires more precise
-    /// hardware like GPS, which consumes more power.
+    /// - note: Never specify an accuracy value greater than what you need.  Core Location uses the
+    ///         accuracy value you specify to manage power better.  A higher degree of accuracy
+    ///         requires more precise hardware like GPS, which consumes more power.
     ///
     /// - parameters:
-    ///   - desiredAccuracy        : The accuracy of the location data.
-    ///   - distanceFilter         : The minimum distance (measured in meters) a device must move
-    ///                              horizontally before an update event is generated.
-    ///   - requestAuthorizeAlways : Whether to requests permission to use location services
-    ///                              whenever the app is running (`true`) or only while the app is
-    ///                              in the foreground (`false`).  Default value is `false`.
+    ///
+    ///   - desiredAccuracy   : The accuracy of the location data.
+    ///   - distanceFilter    : The minimum distance (measured in meters) a device must move
+    ///                         horizontally before an update event is generated.
+    ///   - backgroundUpdates : If set to `true`, will request "always" authorization to receive
+    ///                         location updates whenever the app is running (that is, both when it
+    ///                         is in the foreground and in the background).  If set to `false`
+    ///                         (default), only "while in use" authorization will be requested.
     ///
     /// - seealso:
+    ///
     ///   - [CLLocationManager](apple-reference-documentation://hs8c5staNS#overview)
 
     public static func standardTracker(desiredAccuracy: CLLocationAccuracy,
-                                       distanceFilter: Double,
-                                       requestAuthorizeAlways: Bool = false) -> LocationTracker {
+                                       distanceFilter: Double = kCLDistanceFilterNone,
+                                       backgroundUpdates: Bool = false) -> LocationTracker {
         return StandardLocationTracker(
             desiredAccuracy: desiredAccuracy,
             distanceFilter: distanceFilter,
-            requestAuthorizeAlways: requestAuthorizeAlways
+            backgroundUpdates: backgroundUpdates
         )
     }
 
@@ -55,32 +58,37 @@ extension LocationTracker {
     /// are met, the location manager may begin deferring the delivery of events.
     ///
     /// - parameters:
-    ///   - deferredDistance            : The distance (in meters) from the current location that
-    ///                                   must be travelled before event delivery resumes. To
-    ///                                   specify an unlimited distance, pass the `CLLocationDistanceMax`
-    ///                                   constant.
-    ///   - deferredTimeout             : The amount of time (in seconds) from the current time that
-    ///                                   must pass before event delivery resumes. To specify an
-    ///                                   unlimited amount of time, pass the `CLTimeIntervalMax`
-    ///                                   constant.
-    ///   - extendAccuracyForNavigation : Deferred location updates are only available in the modes
-    ///                                   of `kCLLocationAccuracyBest` or `kCLLocationAccuracyBestForNavigation`.
-    ///                                   Setting this parameters to `true` selects the latter,
-    ///                                   while the `false` (default) opts for the former.
+    ///
+    ///   - deferredDistance      : The distance (in meters) from the current location that
+    ///                             must be travelled before event delivery resumes. To
+    ///                             specify an unlimited distance, pass the `CLLocationDistanceMax`
+    ///                             constant.
+    ///   - deferredTimeout       : The amount of time (in seconds) from the current time that
+    ///                             must pass before event delivery resumes. To specify an
+    ///                             unlimited amount of time, pass the `CLTimeIntervalMax`
+    ///                             constant.
+    ///   - activityType          : The type of user activity associated with the location updates.
+    ///   - accuracyForNavigation : Setting this parameters to `true` selects `kCLLocationAccuracyBestForNavigation`
+    ///                             accuracy for the tracker, while `false` (default) will use
+    ///                            `kCLLocationAccuracyBest`.  These two modes are the only
+    ///                             ones that support deferred updates.
     ///
     /// - seealso:
-    ///   - [CLLocationManager.allowDeferredLocationUpdates(untilTraveled:timeout:)](apple-reference-documentation://hs64cDNHc7)
+    ///
+    ///   - [CLLocationManager.allowDeferredLocationUpdates](untilTraveled:timeout:)](apple-reference-documentation://hs64cDNHc7)
     ///   - [Energy Efficiency Guide for iOS Apps](https://developer.apple.com/library/content/documentation/Performance/Conceptual/EnergyGuide-iOS/LocationBestPractices.html#//apple_ref/doc/uid/TP40015243-CH24-SW8)
-    ///   - [stackoverflow.com](http://stackoverflow.com/a/14509263/1542569)
-    ///   - [stackoverflow.com](http://stackoverflow.com/a/26345001/1542569)
+    ///   - [Answer: How to work with deferred location iOS 6?](http://stackoverflow.com/a/14509263/1542569)
+    ///   - [Answer: ios deferred location updates fail to defer](http://stackoverflow.com/a/26345001/1542569)
 
     public static func deferredLocationUpdatesTracker(deferredDistance: CLLocationDistance,
                                                       deferredTimeout: TimeInterval,
-                                                      extendAccuracyForNavigation: Bool = false) -> LocationTracker {
+                                                      activityType: CLActivityType = CLActivityType.other,
+                                                      accuracyForNavigation: Bool = false) -> LocationTracker {
         return DeferredLocationUpdatesTracker(
             deferredDistance: deferredDistance,
             deferredTimeout: deferredTimeout,
-            extendAccuracyForNavigation: extendAccuracyForNavigation
+            activityType: activityType,
+            accuracyForNavigation: accuracyForNavigation
         )
     }
 
@@ -92,8 +100,8 @@ extension LocationTracker {
     /// standard location services.
     ///
     /// The significant location change service delivers events normally while an app is running in 
-    /// the foreground or background. For a terminated iOS app, this service relaunches the app to 
-    /// deliver events.  Use of this service requires “Always” authorization from the user.
+    /// the foreground or background.  For a terminated iOS app, this service relaunches the app to
+    /// deliver events.  Use of this service requires "Always" authorization from the user.
     ///
     /// - important: 
     ///
@@ -101,11 +109,13 @@ extension LocationTracker {
     ///   considered before significant-change location updates.  In the event significant-change
     ///   location updates are needed, keep in mind the following, which can actually result in
     ///   higher energy use if not employed effectively:
+    ///
     ///   - Significant-change location updates wake the system and your app once every 15 minutes, 
     ///     at minimum, even if no location changes have occurred.
     ///   - Significant-change location updates run continuously, around the clock, until you stop them.
     ///
     /// - seealso:
+    ///
     ///   - [CLLocationManager](apple-reference-documentation://hs8c5staNS#overview)
     ///   - [Energy Efficiency Guide for iOS Apps](https://developer.apple.com/library/content/documentation/Performance/Conceptual/EnergyGuide-iOS/LocationBestPractices.html#//apple_ref/doc/uid/TP40015243-CH24-SW4)